Add files via upload
Customer feedbacks play a vital role in judging the quality of any product or business. It suits the Hotel industry well. here comes the sentiment analysis of a real time dataset about a resort.

Hello All...

This Is Abhinav, Today We Are Going To Take Look About Basic Text Analysis Method.
Using Tools And Using Excel. 
Here We Are Going To Analyse Resort Dataset. It Is Real Time Dataset From A Resort.Where There Is Thousands Of Customers Frequently Visiting For Accomodate.From Different Part Of The World. Different Cultures Of People, And Mainly They Were Representing Different Languages Also. 
In This Data Set There Are Only 3 Columns Available. Reviews Written By Customers, Date Of Received The Review, And Location That The Customer Come From.
From This Junk Of Text , How Could We Can Interpret Insights ! This is The Task That I Done .
From This Project , We Can Understand That, What Is The Most Feedback /Opinion About Hotel From Customers.Is It Possitive Or Negative.
And We Can Understand That Which Word Is Repeated Several Times . 
And We Can Even Find Out How Many Times The Perticular Word Get Repeated.And Also Find Out Which Words Are Secondarily Following The Trend .
And We Can Find Out Which Location People Are Mostly Visited Or Share Their Review .
And I Have To Tell This Too...
Excel Also Plays A Wonderful Role In Text Analysis.
Here ,By Using Excel ,We Can Understand How Much Review Count That Each Location Contributed.And We Can Interpret It By Percentage Formate And Bar Chart Representation.

And Also Finally ,I Will Try To Interpret The Resort Data Using Tableau , And There Also We Get Informative Insights . There We Can Analyse Various Review Written Locations With Their Years (Here Perticularly 2018/2019/2021) .

And Question Is ? What Is The Use ! Is In It... 

We Do Text Analysis For Find Out Or Manipulate The Future Dicision Making. Or Running Business More More Successful Ways. For That Past Customers Reviews Are Very Much Important Data's . So If We Can Intrepret The Past Comments , Then We Can Understand Strength, Weakness, Opportunity And Target Etc.